requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-jose[cryptography]>=3.4.0
tinydb>=4.8.0